change structure of files
